requirements.txt
click==6.6
PyYAML==3.12
lxml==3.6.4
Pillow==4.3.0